refactor(context-menu): H.6 — extract SinglePlaylist + MultiPlaylist submenus
